encryption/requirements.txt
aioredis==1.3.1
amqp==5.1.1
aniso8601==9.0.1
anyio==3.6.2
arabic-reshaper==3.0.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asn1crypto==1.5.1
asttokens==2.2.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.1
Babel==2.11.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
billiard==3.6.4.0
bleach==6.0.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudinary==1.32.0
colorama==0.4.6
comm==0.1.2
constantly==15.1.0
contourpy==1.0.7
cryptography==40.0.2
cssselect2==0.7.0
cycler==0.11.0
daphne==2.5.0
datetime-truncate==1.1.1
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.6
dj-database-url==1.2.0
Django==4.1.4
django-admin-charts==1.2.4
django-admin-interface==0.24.2
django-bower==5.2.0
django-cloudinary-storage==0.3.0
django-colorfield==0.8.0
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-easy-pjax==1.3.0
django-elasticsearch-dsl==7.2.2
django-environ==0.9.0
django-filter==22.1
django-memoize==2.3.1
django-multiselectfield==0.1.12
django-nvd3==0.9.7
django-upgrade==1.12.0
django-utils-six==2.0
django-xhtml2pdf==0.0.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
dnspython==2.2.1
elasticsearch==7.17.8
elasticsearch-dsl==7.4.0
email-validator==1.3.0
entrypoints==0.4
eventlet==0.33.3
exceptiongroup==1.1.0
executing==1.2.0
fastjsonschema==2.16.2
filelock==3.8.0
fonttools==4.38.0
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==2.0.2
gunicorn==20.1.0
h11==0.14.0
hiredis==2.2.3
html5lib==1.1
hyperlink==21.0.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.10.2
incremental==22.10.0
ipykernel==6.20.2
ipython==8.8.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.11
jsonschema==4.17.3
jupyter-client==7.4.9
jupyter-core==5.1.5
jupyter-events==0.6.3
jupyter-server==2.1.0
jupyter-server-terminals==0.4.4
jupyterlab==3.5.3
jupyterlab-pygments==0.2.2
jupyterlab-server==2.19.0
kiwisolver==1.4.4
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mistune==2.0.4
msgpack==1.0.5
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.9
nbformat==5.7.3
nest-asyncio==1.5.6
notebook==6.5.2
notebook-shim==0.2.2
numpy==1.24.1
oscrypto==1.3.0
outcome==1.2.0
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
patsy==0.5.3
pickleshare==0.7.5
Pillow==9.3.0
pkgutil-resolve-name==1.3.10
platformdirs==2.5.4
plotly==5.14.1
portalocker==2.7.0
prometheus-client==0.16.0
promise==2.3
prompt-toolkit==3.0.38
psutil==5.9.4
psycopg2==2.9.5
psycopg2-binary==2.9.5
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.17
pycryptodomex==3.17
Pygments==2.14.0
pyHanko==0.18.1
pyhanko-certvalidator==0.23.0
PyJWT==2.6.0
pymongo==4.3.3
PyMySQL==1.0.2
pyOpenSSL==23.2.0
pyparsing==3.0.9
pypdf==3.9.0
pypiwin32==223
pypng==0.20220715.0
pyrsistent==0.19.3
PySocks==1.7.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.7
python-dotenv==0.21.1
python-json-logger==2.0.4
python-nvd3==0.14.2
python-slugify==1.1.4
python-snappy==0.6.1
pytz==2022.6
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.0
qrcode==7.4.2
redis==4.5.5
reportlab==3.6.13
requests==2.28.2
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
scipy==1.10.0
seaborn==0.11.0
selenium==4.7.2
Send2Trash==1.8.0
service-identity==21.1.0
shellescape==3.8.1
six==1.16.0
sniffio==1.3.0
sockets==1.0.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stack-data==0.6.2
statsmodels==0.13.5
svglib==1.5.1
tenacity==8.2.2
terminado==0.17.1
text-unidecode==1.3
tinycss2==1.2.1
tokenize-rt==5.0.0
tomli==2.0.1
tornado==6.2
traitlets==5.8.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
twisted-iocpsupport==1.0.3
txaio==23.1.1
typing-extensions==4.5.0
tzdata==2022.7
tzlocal==5.0.1
Unidecode==1.3.6
uritools==4.0.1
urllib3==1.26.14
vine==5.0.0
virtualenv==20.17.0
virtualenvwrapper-win==1.2.7
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.4.2
win10toast==0.9
winkerberos==0.9.1
wsproto==1.2.0
xhtml2pdf==0.2.11
zipp==3.11.0
zope.interface==6.0
